ref: moved Article schemas to /schemas/article
