+++ schemas/article/Article.ts
